requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
httpx==0.27.0
redis==5.0.4
psycopg2-binary==2.9.9
pydantic==2.7.1
slowapi==0.1.7
tenacity==8.2.3
prometheus_fastapi_instrumentator==6.1.0
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
